# allow INST to be used dynamically