@stylewarning 2020-03-30
(May be of interest to @ecpeterson.)