@stylewarning 2020-03-30
For N = 1 samples, we are actually 10 seconds _faster_.

before:
```
Evaluation took:
  462.272 seconds of real time
  463.079237 seconds of total run time (460.666083 user, 2.413154 system)
  [ Run times consist of 3.633 seconds GC time, and 459.447 seconds non-GC time. ]
  100.17% CPU
  92,424 forms interpreted
  13,567 lambdas converted
  1,109,454,218,200 processor cycles
  323,737,789,280 bytes consed
```

after:
```
  451.827 seconds of real time
  452.877278 seconds of total run time (450.658180 user, 2.219098 system)
  [ Run times consist of 3.461 seconds GC time, and 449.417 seconds non-GC time. ]
  100.23% CPU
  92,424 forms interpreted
  4,796 lambdas converted
  1,084,334,105,504 processor cycles
  320,138,089,840 bytes consed
```